Further updates to reacting algorithm
Previously we introduced collapsing a string of consecutive characters
for reacting purposes. Now we add bounds to that -- a minimum bound of
repeated characters that must occur for collapsing to occur, and a
maximum bound that must not be exceeded in the reaction itself.

When reacting, we require min=3, and try both max=2 and max=1 to find an
appropriate fallback reaction.